+++ addons/yat/src/scenes/debug_screen/DebugScreen.cs
@@ -1,0 +1,9 @@
+namespace YAT.Scenes;
+
+public partial class DebugScreen : YatWindow
+{
+	public override void _Ready()
+	{
+		base._Ready();
+	}
+}