+++ CHANGELOG.md
@@ -2,6 +2,12 @@
 
 All notable changes to this project will be documented in this file.
 
+## [Unreleased]
+
+### Added
+
+-   DebugScreen scene.
+
 ## [1.24.1-beta 2024-02-24]
 
 ### Added